requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gradio==4.44.1
SQLAlchemy==2.0.25
deepseek==1.0.0
pytest==7.4.3
black==23.12.1
flake8==7.0.0
python-jose==3.3.0
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.3
gunicorn==21.2.0
FastAPI==0.110.0
uvicorn==0.27.1
email-validator==2.1.1
PyJWT==2.8.0
httpx==0.27.0
pydantic[email]